requirements.txt
asgiref==3.8.1
async-property==0.2.2
chardet==5.2.0
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-admin-rangefilter==0.13.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.53.1
gunicorn==22.0.0
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
reportlab==4.2.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
